Git Notes A lot of these are probably bugs because we couldn’t actually test test it so we made mistakes and didn’t catch them.  I’m putting everything here from the enviornmnet in case I changed something I shouldn’t have.
Added preconditions to all functions the return true

Line 238 in draw_env: Changed return self._obesrve() to self.get_observation()
Moved self.has_been_reset = True above the return

In drawenv I changed current pixel data to always be a numpy array.  It was a pixel access object before which I don’t think was the intention.  I’ll probably branch this tonight before pushing so I don’t break everything.

In draw_env DrawEnvEncoder forward added line x = x.view(-1,1,200,200) to get the flattened data to fit the convolutional layer and removed the dilation=0 in __init__ because a convoution apparently won’t take 0.

The encoding dim in the DrawEnvEncoder is setting the number of output covolutions so we end up with the encoder outputting 32 200x200 canvases.  I’m added the max pooling and changed the output layer to a densenet with output dim outputs.  I don’t know if that will work but I can debug a lot of the rest of it and we can change it later If we need to.
I significantly changed the encoding network so we should probably look at the parameters once it’s running.

Changed the activation function for the beta output of ContinuousActorNet from ReLu to SoftPlus to force it positive.  Sampling from the numpy beta distribution doens’t take zero as a valid input.

The in _move in draw_env,   line  (and other shape) fucntions from skimage/draw only take ints as inputs.  I hacked something with np.linspace so I could keep debugging but we’re going to have to have something real there later.  I think you said you had something you had written already that you replaced so I just did something I could write quickly.

I changed the post condition for line in the same way to generate the drawing.  I also changed line to return the target canvas as well as if it was complete.  I needed the target for get_reward() to give a reward that the value network uses for loss.  I don’t think this will break anything but it seems inefficient to make the target board again and again.

I changed self.stride in draw_env to 1 as I think it is how far the move command is moving each call.

Removed the self arguments from the postcondition functions that take arguments

I had widened nodes take the beta parameters from their parents because I think they should still be good for initial values as rewards are moving back up the rollout.  I need to go back and check if I need to do a forward pass to get different parameters.

I put in some code in move to keep it from going off the grid.  Right now it just puts it back where it was so it should learn going off the edge doesn’t increase reward but we may want to add a penalty if it keeps trying to move in that directions